requirements.txt
auth==0.5.3
bcrypt==3.1.7
blinker==1.4
boto==2.49.0
boto3==1.10.20
botocore==1.13.20
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Django==2.2.1
django-crispy-forms==1.8.0
django-forms-bootstrap==3.1.0
dnspython==1.16.0
docutils==0.15.2
envy==0.2.0
eventlet==0.25.1
falcon==2.0.0
Flask==1.1.1
greenlet==0.4.15
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
itsdangerous==1.1.0
jmespath==0.9.4
mechanize==0.4.4
mongoengine==0.18.2
monotonic==1.5
mysql==0.0.2
mysqlclient==1.4.5
paramiko==2.6.0
Pillow==6.0.0
pycparser==2.19
pymongo==3.9.0
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
SQLAlchemy==1.3.4
sqlparse==0.3.0
storage==0.0.4.3
stripe==2.40.0
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
